Edited activityNumberOfWeek type
Changed return type of ActivityIfc.getActivityNumberOfWeek from int to
byte (which is used in the implementation Activity)

+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/simulation/activityschedule/Activity.java
@@ -156,7 +156,7 @@ public class Activity implements ActivityIfc {
     return this.mode;
   }
 
-  public int getActivityNrOfWeek() {
+  public byte getActivityNrOfWeek() {
   	return activityNrOfWeek;
   }
